Move everything to AwsAuroraPgvectorServerlessNestedStack stack
